personal_tracker/requirements.txt
# Flask and web dependencies
Flask==2.2.3
Flask-WTF==1.1.1
Werkzeug==2.2.3

# OAuth and API dependencies
requests==2.28.2
oauthlib==3.2.2
requests-oauthlib==1.3.1
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-api-python-client==2.80.0

# Security
cryptography==39.0.2
PyJWT==2.6.0
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
APScheduler==3.10.1

# Testing
pytest==7.3.1
pytest-flask==1.2.0
